Fixing prechecking failures
1) Ignore Literal while converting an expression to DynamicFoldable
because we are interested for ParamLiterals only and Literal past
parsing gets introduced only via rules or api.

2) Moved the DFExp class outside SnappySessionState parent class
avoiding serialization exceptions.
